Adding support for "login as user" to the venue manager. [touch:378]

+++ includes/admin-files/venue-management/update_venue.php
@@ -1,6 +1,6 @@
 <?php 
 function update_event_venue(){
-	global $wpdb;
+	global $wpdb, $notices;
 	$wpdb->show_errors();
 	//print_r($_REQUEST);
 	$venue_meta['contact'] = $_REQUEST['contact'];
@@ -15,30 +15,36 @@ function update_event_venue(){
 	$meta = serialize($venue_meta);
 		
 	
-	$sql=array('name'=>$_REQUEST['name'],'address'=>$_REQUEST['address'], 'address2'=>$_REQUEST['address2'], 'city'=>$_REQUEST['city'], 'state'=>$_REQUEST['state'], 'zip'=>$_REQUEST['zip'], 'country'=>$_REQUEST['country'], 'meta'=>$meta); 
+	$sql=array(
+		'name'=>$_REQUEST['name'],
+		'address'=>$_REQUEST['address'],
+		'address2'=>$_REQUEST['address2'],
+		'city'=>$_REQUEST['city'],
+		'state'=>$_REQUEST['state'],
+		'zip'=>$_REQUEST['zip'],
+		'country'=>$_REQUEST['country'],
+		'meta'=>$meta); 
 		
 	$update_id = array('id'=> $_REQUEST['venue_id']);
-	/*echo 'Debug: <br />';
-	print_r($sql);
-	print 'Number of vars: ' . count ($sql);
-	echo '<br />';
-	print 'Number of cols: ' . count($sql_data);*/
-		
 	$sql_data = array('%s','%s','%s','%s','%s','%s','%s','%s');
-	$wpdb->update( EVENTS_VENUE_TABLE, $sql, $update_id, $sql_data, array( '%d' ) );
 	
-			if( !empty($locale) ){
-				$last_venue_id = $_REQUEST['venue_id'];
-				$wpdb->query("DELETE FROM ".EVENTS_LOCALE_REL_TABLE." WHERE venue_id='".$last_venue_id."'");
-				$sql_locale="INSERT INTO ".EVENTS_LOCALE_REL_TABLE." (venue_id, locale_id) VALUES ('".$last_venue_id."', '".$locale."')";
-				if (!$wpdb->query($sql_locale)){
-					$error = true;
-				}
+	if ($wpdb->update( EVENTS_VENUE_TABLE, $sql, $update_id, $sql_data, array( '%d' )) ){
+		/*echo 'Debug: <br />';
+		print_r($sql);
+		print 'Number of vars: ' . count ($sql);
+		echo '<br />';
+		print 'Number of cols: ' . count($sql_data);*/
+		
+		if( !empty($locale) ){
+			$last_venue_id = $_REQUEST['venue_id'];
+			$wpdb->query("DELETE FROM ".EVENTS_LOCALE_REL_TABLE." WHERE venue_id='".$last_venue_id."'");
+			$sql_locale="INSERT INTO ".EVENTS_LOCALE_REL_TABLE." (venue_id, locale_id) VALUES ('".$last_venue_id."', '".$locale."')";
+			if (!$wpdb->query($sql_locale)){
+				$notices['errors'][] = __('The locale was not saved!', 'event_espresso');
 			}
-?>	
-	<div id="message" class="updated fade">
-			<p><strong><?php _e('The venue  has been updated.','event_espresso'); ?></strong></p>
-		 </div>	
-		 
-<?php
+		}
+		$notices['updates'][] = __('The venue ', 'event_espresso') . $_REQUEST['name'] .  __(' has been added', 'event_espresso');
+	}else{ 
+		$notices['errors'][] = __('The venue', 'event_espresso') . $_REQUEST['name'] .  __(' was not saved!', 'event_espresso');		
+	}
 }
